+++ src/components/Editor.js
@@ -6,7 +6,6 @@ import Fab from '@material-ui/core/Fab';
 import AddIcon from '@material-ui/icons/Add';
 import DeleteIcon from '@material-ui/icons/Delete';
 import Paper from '@material-ui/core/Paper'
-import axios from 'axios';
 
 const styles = theme => ({
   fab: {
@@ -21,24 +20,25 @@ const styles = theme => ({
 });
 
 class Editor extends React.Component {
-  state = { text: '' }
+  state = {
+    title: 'test',
+    body: '' ,
+    creator_id: '1'
+  }
   handleChange = (value) => {
-    this.setState({ text: value });
+    this.setState({ body: value });
   }
 
   render() {
     return (
       <div>
       <Paper>
-      <ReactQuill value={this.state.text} onChange={this.handleChange} />
+      <ReactQuill value={this.state.body} onChange={this.handleChange} />
       </Paper>
-      <Fab color="secondary" aria-label="Delete" onClick={() => { console.log(`Delete from store`) }}>
+      <Fab color="secondary" aria-label="Delete" onClick={() => { console.log('delete') }} >
         <DeleteIcon />
       </Fab>
-      <Fab color="primary" aria-label="Axios Post" onClick={
-        () => { 
-      axios.post('https://yanna-howto.herokuapp.com/api/posts/', this.state.text) }
-      }>
+      <Fab color="primary" aria-label="Axios Post" onClick={() => { console.log('post') }} >
         <AddIcon />
       </Fab>
     </div>
